change reward 0.5

+++ src/leveldb/db/memtable.cc
@@ -101,7 +101,7 @@ void MemTable::Add(SequenceNumber s, ValueType type,
   p += 8;
   p = EncodeVarint32(p, val_size);
   memcpy(p, value.data(), val_size);
-  assert((size_t)((p + val_size) - buf) == encoded_len);
+  assert((p + val_size) - buf == encoded_len);
   table_.Insert(buf);
 }
 
@@ -113,7 +113,7 @@ bool MemTable::Get(const LookupKey& key, std::string* value, Status* s) {
     // entry format is:
     //    klength  varint32
     //    userkey  char[klength]
-    //    tag      uint64_t
+    //    tag      uint64
     //    vlength  varint32
     //    value    char[vlength]
     // Check that it belongs to same user key.  We do not check the
